# api/docs/openapi.yaml
openapi: 3.0.3
paths:
...
  /api/v1/users/me/categories:
    post:
      description: Category names are trimmed before storage and uniqueness checking; uniqueness is case-insensitive.
      operationId: Categories_createCategory
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: Category created.
          headers:
            Location:
              required: true
              description: Relative canonical URI of the created Category.
              schema:
                type: string
                example: /api/v1/users/me/categories/42
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '413':
          $ref: '#/components/responses/HttpError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaTypeError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Create a category.
      tags:
        - Categories
    get:
      operationId: Categories_listCategories
      parameters: []
      responses:
        '200':
          description: Categories.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponseDto'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: List all owned categories in ascending ID order.
      tags:
        - Categories
  /api/v1/users/me/categories/{categoryId}:
    get:
      operationId: Categories_getCategory
      parameters:
        - name: categoryId
          required: true
          in: path
          description: Positive bigint identifier encoded as a decimal JSON string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
      responses:
        '200':
          description: Category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Get a category.
      tags:
        - Categories
    patch:
      description: Inactive Categories retain their historical relationships and cannot receive a new Budget, although an existing Budget may still be replaced.
      operationId: Categories_updateCategory
      parameters:
        - name: categoryId
          required: true
          in: path
          description: Positive bigint identifier encoded as a decimal JSON string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: Category updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '413':
          $ref: '#/components/responses/HttpError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaTypeError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Edit or activate/deactivate a category.
      tags:
        - Categories
...
  /api/v1/users/me/category-summaries:
    get:
      description: Active categories are included even with no spending; inactive categories remain only when they have spending, and results are ordered by category ID. The month requirement depends on period and is enforced at runtime because OpenAPI cannot express that cross-field rule. Unknown query parameters are rejected.
      operationId: CategorySummaries_getCategorySummary
      parameters:
        - name: period
          required: true
          in: query
          description: Calendar period. Monthly summaries require month and yearly summaries do not accept month.
          schema:
            example: monthly
            type: string
            enum:
              - monthly
              - yearly
        - name: year
          required: true
          in: query
          description: Four-digit calendar year used for the summary.
          schema:
            pattern: ^\d{4}$
            example: '2026'
            type: string
        - name: month
          required: false
          in: query
          description: Required for monthly summaries; forbidden for yearly summaries.
          schema:
            pattern: ^(0[1-9]|1[0-2])$
            example: '08'
            type: string
      responses:
        '200':
          description: Category summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategorySummaryResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Get a complete monthly or yearly calendar-period summary.
      tags:
        - Category summaries
...
  /api/v1/users/me/spaces/{spaceId}/categories/{categoryId}:
    get:
      operationId: SpaceCategories_getCategory
      parameters:
        - name: categoryId
          required: true
          in: path
          description: Positive bigint Category identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
        - name: spaceId
          required: true
          in: path
          description: Positive bigint Space identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '7'
      responses:
        '200':
          description: Category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Get a Category in an authorized Space.
      tags:
        - Categories
    patch:
      operationId: SpaceCategories_updateCategory
      parameters:
        - name: categoryId
          required: true
          in: path
          description: Positive bigint Category identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '42'
        - name: spaceId
          required: true
          in: path
          description: Positive bigint Space identifier encoded as a string.
          schema:
            type: string
            pattern: ^[1-9]\d*$
            example: '7'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: Category updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UserNotProvisionedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '413':
          $ref: '#/components/responses/HttpError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaTypeError'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Edit or activate/deactivate a Category in an authorized Space.
      tags:
        - Categories
...
info:
  title: Expense Tracker REST API
  description: A user-scoped expense tracker API. Bigint identifiers and counts, decimal money, domain dates, and UTC timestamps use the encodings described by the schemas below.
  version: 1.0.0
  contact: {}
...
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        code:
          description: Stable error code. Human-readable messages are not a compatibility contract.
          allOf:
            - $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Human-readable error message.
        details:
          type: array
          description: Structured details for fields or domain-specific conflicts.
          items:
            oneOf:
              - $ref: '#/components/schemas/ErrorDetailDto'
              - $ref: '#/components/schemas/ProbableDuplicateDetailDto'
      additionalProperties: false
      required:
        - code
        - message
        - details
    ErrorEnvelopeDto:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorResponseDto'
      additionalProperties: false
      required:
        - error
...
    ErrorDetailDto:
      type: object
      properties:
        categoryId:
          type: string
          description: Conflicting owned Category identifier for a category-rule duplicate.
          pattern: ^[1-9]\d*$
          example: '42'
        field:
          type: string
          description: JSON Pointer into the request.
          example: /email
        code:
          description: Stable machine-readable field-detail code.
          allOf:
            - $ref: '#/components/schemas/ErrorDetailCode'
        message:
          type: string
          description: Human-readable detail message.
      additionalProperties: false
      required:
        - field
        - code
        - message
...
    CategoryResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Positive bigint Category identifier encoded as a string.
          pattern: ^[1-9]\d*$
          example: '42'
        name:
          type: string
          description: Trimmed Category display name. Uniqueness is checked case-insensitively.
          minLength: 1
          maxLength: 100
          pattern: \S
          example: Dining Out
        description:
          type: string
          description: Trimmed Category description, or null when no description is set.
          nullable: true
          maxLength: 500
          example: Restaurants and cafes
        color:
          type: string
          enum:
            - coral
            - scarlet
            - crimson
            - rose
            - magenta
            - orchid
            - plum
            - violet
            - indigo
            - cobalt
            - azure
            - sky
            - cyan
            - teal
            - emerald
            - forest
            - lime
            - olive
            - gold
            - amber
            - copper
            - cocoa
            - slate
            - steel
          description: Effective Category Color palette identifier. An explicitly saved choice is returned as-is; Categories without a saved choice receive a stable identity-based default.
          example: teal
        isActive:
          type: boolean
          description: Whether the Category is active. Inactive Categories retain their historical relationships and cannot receive a new Budget, although an existing Budget may still be replaced.
          example: true
        createdAt:
          type: string
          description: UTC timestamp when the Category was created.
          format: date-time
          example: '2026-08-29T00:00:00.000Z'
        updatedAt:
          type: string
          description: UTC timestamp when the Category was last updated.
          format: date-time
          example: '2026-08-29T00:00:00.000Z'
      description: A Category owned by the authenticated User. Inactive Categories retain their historical relationships.
      additionalProperties: false
      required:
        - id
        - name
        - description
        - color
        - isActive
        - createdAt
        - updatedAt
    CreateCategoryDto:
      type: object
      properties:
        name:
          type: string
          description: Category display name. It is trimmed before storage and uniqueness checking; uniqueness is case-insensitive.
          minLength: 1
          maxLength: 100
          pattern: \S
          example: Dining Out
        description:
          type: string
          description: Optional Category description. Leading and trailing whitespace is removed; blank text is stored as null.
          nullable: true
          maxLength: 500
          example: Restaurants and cafes
        color:
          type: string
          enum:
            - coral
            - scarlet
            - crimson
            - rose
            - magenta
            - orchid
            - plum
            - violet
            - indigo
            - cobalt
            - azure
            - sky
            - cyan
            - teal
            - emerald
            - forest
            - lime
            - olive
            - gold
            - amber
            - copper
            - cocoa
            - slate
            - steel
          description: Optional palette identifier for the Category Color. When omitted, the API resolves a stable default from the Category identity.
          example: teal
      additionalProperties: false
      required:
        - name
    UpdateCategoryDto:
      type: object
      properties:
        name:
          type: string
          description: Category display name. It is trimmed before storage and uniqueness checking; uniqueness is case-insensitive.
          minLength: 1
          maxLength: 100
          pattern: \S
          example: Dining Out
        description:
          type: string
          description: Optional Category description. Leading and trailing whitespace is removed; null or blank text clears it.
          nullable: true
          maxLength: 500
          example: Restaurants and cafes
        isActive:
          type: boolean
          description: Whether the Category is active. Inactive Categories retain their historical relationships and cannot receive a new Budget, although an existing Budget may still be replaced.
          example: false
        color:
          type: string
          enum:
            - coral
            - scarlet
            - crimson
            - rose
            - magenta
            - orchid
            - plum
            - violet
            - indigo
            - cobalt
            - azure
            - sky
            - cyan
            - teal
            - emerald
            - forest
            - lime
            - olive
            - gold
            - amber
            - copper
            - cocoa
            - slate
            - steel
          description: Palette identifier for the Category Color. Saved colors remain unchanged when this field is omitted.
          example: teal
        updatedAt:
          type: string
          description: The UTC timestamp returned by the last read. Stale edits are rejected when another save changed the Category first.
          format: date-time
          example: '2026-08-29T00:00:00.000Z'
      additionalProperties: false
      minProperties: 1
...
    CategorySummaryResponseDto:
      type: object
      properties:
        period:
          type: string
          description: Requested calendar period.
          enum:
            - monthly
            - yearly
          example: monthly
        year:
          type: string
          description: Four-digit calendar year used for the summary.
          pattern: ^\d{4}$
          example: '2026'
        month:
          type: string
          description: Two-digit calendar month for a monthly summary, or null for a yearly summary.
          pattern: ^(0[1-9]|1[0-2])$
          nullable: true
          example: '08'
        categories:
          type: array
          description: Categories ordered by ascending Category ID. Active Categories are included even with no spending; inactive Categories are included only when they have spending in the requested period.
          items:
            $ref: '#/components/schemas/CategorySummaryItemResponseDto'
        uncategorizedTotal:
          type: string
          description: Non-negative exact two-decimal spending total for Uncategorized Transactions, encoded as a string and zero when none exist. Aggregate totals are not limited by per-Transaction numeric precision.
          pattern: ^\d+\.\d{2}$
          example: '4.50'
        uncategorizedCount:
          type: string
          description: Non-negative count of Uncategorized Transactions encoded as a string.
          pattern: ^\d+$
          example: '1'
      description: A complete monthly or yearly Category spending summary owned by the authenticated User.
      additionalProperties: false
      required:
        - period
        - year
        - month
        - categories
        - uncategorizedTotal
        - uncategorizedCount
    CategorySummaryItemResponseDto:
      type: object
      properties:
        categoryId:
          type: string
          description: Positive bigint Category identifier encoded as a string.
          pattern: ^[1-9]\d*$
          example: '42'
        name:
          type: string
          description: Trimmed Category display name.
          minLength: 1
          maxLength: 100
          pattern: \S
          example: Groceries
        isActive:
          type: boolean
          description: Whether the Category is active.
          example: true
        totalAmount:
          type: string
          description: Non-negative exact two-decimal aggregate spending amount encoded as a string. Aggregate totals are not limited by per-Transaction numeric precision.
          pattern: ^\d+\.\d{2}$
          example: '1250.00'
        transactionCount:
          type: string
          description: Non-negative Transaction count encoded as a string.
          pattern: ^\d+$
          example: '12'
        budgetAmount:
          type: string
          description: Matching-period Budget amount encoded as a positive exact two-decimal string, or null when the Category has no Budget for the requested period.
          pattern: ^(?=.*[1-9])\d{1,13}\.\d{2}$
          nullable: true
          example: '1500.00'
        remainingAmount:
          type: string
          description: Budget amount minus aggregate spending, encoded as an exact two-decimal string; it may be negative when spending exceeds the matching-period Budget, and is null when budgetAmount is null.
          pattern: ^-?\d+\.\d{2}$
          nullable: true
          example: '-25.50'
      description: A Category aggregate for the requested calendar period. Inactive Categories are included only when they have spending in that period.
      additionalProperties: false
      required:
        - categoryId
        - name
        - isActive
        - totalAmount
        - transactionCount
        - budgetAmount
        - remainingAmount
...
  responses:
    UnauthenticatedError:
      description: The request does not include valid Clerk session credentials.
      headers:
        WWW-Authenticate:
          required: true
          description: Authentication scheme accepted by the API.
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: UNAUTHENTICATED
              message: Authentication required
              details: []
    UserNotProvisionedError:
      description: The verified Clerk identity is not linked to a local User. Provision the User first.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: USER_NOT_PROVISIONED
              message: User is not provisioned
              details: []
    NotFoundError:
      description: The requested resource is absent or owned by another user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          examples:
            user:
              value:
                error:
                  code: USER_NOT_FOUND
                  message: User was not found
                  details: []
            route:
              value:
                error:
                  code: ROUTE_NOT_FOUND
                  message: Request failed
                  details: []
    NotAcceptableError:
      description: The client did not accept the representation required by the endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: NOT_ACCEPTABLE
              message: The endpoint requires an application/json response
              details: []
    InternalError:
      description: An unexpected server error. Infrastructure details are never returned.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: INTERNAL_ERROR
              message: An internal error occurred
              details: []
    ValidationError:
      description: The request syntax or fields are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          examples:
            invalidField:
              value:
                error:
                  code: VALIDATION_FAILED
                  message: The request contains invalid fields.
                  details:
                    - field: /email
                      code: invalid_format
                      message: Email must be a valid email address
    ConflictError:
      description: The request conflicts with current application state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          examples:
            emailAlreadyExists:
              value:
                error:
                  code: EMAIL_ALREADY_EXISTS
                  message: Email is already in use
                  details:
                    - field: /email
                      code: not_unique
                      message: Email is already in use
            genericConflict:
              value:
                error:
                  code: CONFLICT
                  message: Request conflicts with existing data
                  details: []
...
    HttpError:
      description: A generic HTTP protocol error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: HTTP_ERROR
              message: Request failed
              details: []
    UnsupportedMediaTypeError:
      description: A request body must use application/json.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeDto'
          example:
            error:
              code: UNSUPPORTED_MEDIA_TYPE
              message: Request bodies must use application/json
              details: []
...
security:
  - bearerAuth: []